LayoutTests/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-mode.any-expected.txt

CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=301&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=302&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=303&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=307&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: No-Cors mode requires follow redirect mode
CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/redirect.py?redirect_status=308&location=cors-top.txt due to access control checks.
CONSOLE MESSAGE: Cross-origin redirection to http://web-platform.test:8800/common/blank.html denied by Cross-Origin Resource Sharing policy: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 302
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/common/redirect.py?location=http://web-platform.test:8800/common/blank.html due to access control checks.

PASS same-origin redirect 301 in error redirect and cors mode
PASS same-origin redirect 301 in error redirect and no-cors mode
PASS same-origin redirect 301 in manual redirect and cors mode
PASS same-origin redirect 301 in manual redirect and no-cors mode
PASS same-origin redirect 301 in follow redirect and cors mode
PASS same-origin redirect 301 in follow redirect and no-cors mode
PASS same-origin redirect 302 in error redirect and cors mode
PASS same-origin redirect 302 in error redirect and no-cors mode
PASS same-origin redirect 302 in manual redirect and cors mode
PASS same-origin redirect 302 in manual redirect and no-cors mode
PASS same-origin redirect 302 in follow redirect and cors mode
PASS same-origin redirect 302 in follow redirect and no-cors mode
PASS same-origin redirect 303 in error redirect and cors mode
PASS same-origin redirect 303 in error redirect and no-cors mode
PASS same-origin redirect 303 in manual redirect and cors mode
PASS same-origin redirect 303 in manual redirect and no-cors mode
PASS same-origin redirect 303 in follow redirect and cors mode
PASS same-origin redirect 303 in follow redirect and no-cors mode
PASS same-origin redirect 307 in error redirect and cors mode
PASS same-origin redirect 307 in error redirect and no-cors mode
PASS same-origin redirect 307 in manual redirect and cors mode
PASS same-origin redirect 307 in manual redirect and no-cors mode
PASS same-origin redirect 307 in follow redirect and cors mode
PASS same-origin redirect 307 in follow redirect and no-cors mode
PASS same-origin redirect 308 in error redirect and cors mode
PASS same-origin redirect 308 in error redirect and no-cors mode
PASS same-origin redirect 308 in manual redirect and cors mode
PASS same-origin redirect 308 in manual redirect and no-cors mode
PASS same-origin redirect 308 in follow redirect and cors mode
PASS same-origin redirect 308 in follow redirect and no-cors mode
PASS cross-origin redirect 301 in error redirect and cors mode
PASS cross-origin redirect 301 in error redirect and no-cors mode
PASS cross-origin redirect 301 in manual redirect and cors mode
PASS cross-origin redirect 301 in manual redirect and no-cors mode
PASS cross-origin redirect 301 in follow redirect and cors mode
PASS cross-origin redirect 301 in follow redirect and no-cors mode
PASS cross-origin redirect 302 in error redirect and cors mode
PASS cross-origin redirect 302 in error redirect and no-cors mode
PASS cross-origin redirect 302 in manual redirect and cors mode
PASS cross-origin redirect 302 in manual redirect and no-cors mode
PASS cross-origin redirect 302 in follow redirect and cors mode
PASS cross-origin redirect 302 in follow redirect and no-cors mode
PASS cross-origin redirect 303 in error redirect and cors mode
PASS cross-origin redirect 303 in error redirect and no-cors mode
PASS cross-origin redirect 303 in manual redirect and cors mode
PASS cross-origin redirect 303 in manual redirect and no-cors mode
PASS cross-origin redirect 303 in follow redirect and cors mode
PASS cross-origin redirect 303 in follow redirect and no-cors mode
PASS cross-origin redirect 307 in error redirect and cors mode
PASS cross-origin redirect 307 in error redirect and no-cors mode
PASS cross-origin redirect 307 in manual redirect and cors mode
PASS cross-origin redirect 307 in manual redirect and no-cors mode
PASS cross-origin redirect 307 in follow redirect and cors mode
PASS cross-origin redirect 307 in follow redirect and no-cors mode
PASS cross-origin redirect 308 in error redirect and cors mode
PASS cross-origin redirect 308 in error redirect and no-cors mode
PASS cross-origin redirect 308 in manual redirect and cors mode
PASS cross-origin redirect 308 in manual redirect and no-cors mode
PASS cross-origin redirect 308 in follow redirect and cors mode
PASS cross-origin redirect 308 in follow redirect and no-cors mode
PASS manual redirect with a CORS error should be rejected
